Add generic OAuth boilerplate
