+++ .env.example
@@ -25,11 +25,20 @@ TECHNICAL_ACCOUNT_TOKEN=
 TECHNICAL_ACCOUNT_USERNAME="weiss-bot"
 TECHNICAL_ACCOUNT_EMAIL="weiss-bot@example.com"
 
-# Microsoft Auth
-MS_AUTH_CLIENT_ID="your-ms-client-id-here"
-MS_AUTH_TENANT_ID="your-ms-tenant-id-here"
-MS_AUTH_CLIENT_SECRET="your-ms-client-secret-here"
+# Auth Client
+AUTH_CLIENT_ID="your-client-id-here"
+# Only used for Microsoft auth
+AUTH_TENANT_ID="your-tenant-id-here"
+AUTH_CLIENT_SECRET="your-client-secret-here"
+
+# Generic OAuth2 configuration
+AUTH_ISSUER="http://localhost:8089/realms/master"
+AUTH_IDENTITY_PROVIDER="oauth"
 
 # Documentation
 # uncomment the following serve docs (requires docs/docker-compose.yml to be running)
 # DOCS_HOSTNAME="docs.example.com" 
+
+# Ports
+HTTP_PORT=80
+HTTPS_PORT=443
