+++ .gitignore
@@ -255,3 +255,5 @@ __marimo__/
 
 # Streamlit
 .streamlit/secrets.toml
+
+backend/api/src/api/_version.py
