+++ docker-compose.yml
@@ -56,9 +56,9 @@ services:
       - "8000:8000"
     restart: always
     environment:
-      MS_AUTH_CLIENT_ID: ${MS_AUTH_CLIENT_ID}
-      MS_AUTH_TENANT_ID: ${MS_AUTH_TENANT_ID}
-      MS_AUTH_CLIENT_SECRET: ${MS_AUTH_CLIENT_SECRET}
+      AUTH_CLIENT_ID: ${AUTH_CLIENT_ID}
+      AUTH_TENANT_ID: ${AUTH_TENANT_ID}
+      AUTH_CLIENT_SECRET: ${AUTH_CLIENT_SECRET}
       APP_HOSTNAME: ${APP_HOSTNAME}
       ENABLE_HTTPS: ${ENABLE_HTTPS:-false}
       TECHNICAL_ACCOUNT_TOKEN: ${TECHNICAL_ACCOUNT_TOKEN}
